Fix initialization of Expires to use DateTimeOffset.MaxValue
Logic was using DateTime.MaxValue which could establish a value that is later adjusted for time zone.  DateTimeOffset.MaxValue is already set to UTC which is the default for DateTimeOffset.

Additional problem was in Engine.cs - there are a couple of of commands that are used in the test harness that need to be declared as public so they can be tested.

Added some comments in PortSerialManager.cs about the possibility of eating non-serial problems in the try/catch block.

Fix #742