Local locks for global NanoFrameworkDevices collection because of discovery of devices is asynchronous (and because of the use of multiple async PortManagers in the test platform v3): - Make sure all list operations use a lock of the collection - PortTcpIpManager._networkDevices made static as it acts as a complement to the NanoFrameworkDevices collection - if there are multiple, the device list management does not work properly for a non-static _networkDevices.
System-wide exclusive access to a device via system-wide mutex, required to avoid competition between e.g.,Device Explorer auto-discovery, test platform etc. If this is used, it is possible for the nanoFramework tools to wait for a device to become available instead of throwing an exception. If an exception is thrown, that is an indication of a real problem, not a competition with other nanoFramework tools. Not sure if that also means the WaitAndRetry in the PortSerialManager should be removed.

Use of System-wide exclusive access in the serial DeviceWatcher; async part of that operation moved from PortSerialManager to DeviceWatcher. Also updated PortTcpIpManager+DeviceWatcher.

Serial ports: DeviceWatcher now has a static method to find all serial ports, incl exclusion list, so that all nanoFramework devices use the same method. Also to be used in the test platform v3.

Port**Manager returns the added or already registered device for PostBase.AddDevice. Makes use in test platform code easier.

Fixed OnDeviceEnumerationCompleted logic if both AddDevice and a device watcher are used. The OnDeviceEnumerationCompleted is now (indirectly) fired by the DeviceWatcher.

Also fixed at least one statement that would result in an exception if NanoNetworkDevices have been discovered before NanoSerialDevices.

Signed-off-by: Frank Robijn <robijn@good-heavens.nl>